fix: date range picker improvement
