+++ apps/gauzy/src/app/@theme/components/header/selectors/date-range-picker/date-picker.utils.ts
@@ -1,14 +1,6 @@
 import * as moment from 'moment';
 import { IDateRangePicker, ISelectedDateRange, ITimeLogFilters, WeekDaysEnum } from "@gauzy/contracts";
-
-export enum DateRangeKeyEnum {
-	TODAY = 'Today',
-	YESTERDAY = 'Yesterday',
-	CURRENT_WEEK = 'Current week',
-	LAST_WEEK = 'Last week',
-	CURRENT_MONTH = 'Current month',
-	LAST_MONTH = 'Last month'
-}
+import { TimePeriod } from './date-picker.interface';
 
 /**
  * We are having issue, when organization not allowed future date
@@ -45,6 +37,24 @@ export function getAdjustDateRangeFutureAllowed(request: ITimeLogFilters | IDate
         startDate: moment(startDate).toDate(),
         endDate: moment(endDate).toDate()
     } as ISelectedDateRange
+}
+
+/**
+ * Shifts a given time range from UTC to the local time zone.
+ *
+ * @param range The time range to be shifted.
+ * @returns The shifted time range in the local time zone.
+ */
+export function shiftUTCtoLocal(range: TimePeriod): TimePeriod {
+    if (range && range.endDate && range.startDate) {
+        const offset = moment().utcOffset();
+        return {
+            startDate: moment(range.startDate.toDate()).subtract(offset, 'minute'),
+            endDate: moment(range.endDate.toDate()).subtract(offset, 'minute'),
+        };
+    } else {
+        return range;
+    }
 }
 
 /**
