+++ apps/gauzy/src/app/@theme/components/header/selectors/date-range-picker/index.ts
@@ -1,4 +1,4 @@
 export * from './date-range-picker.component';
 export * from './date-range-picker.module';
-export * from './date-range-picker.setting';
+export * from './date-picker.utils';
 export * from './date-range-picker.resolver';
